+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog], and this project adheres to
 - Config file check before getopts.
 - Config file read before getopts.
 
+### Updated
+
+- README with configuration information.
+
 ## [0.2.0] - 2021-01-09
 
 ### Added