+++ CHANGELOG.md
@@ -51,6 +51,7 @@ need to address.
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/inigochoa/porg/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/inigochoa/porg/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/inigochoa/porg/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/inigochoa/porg/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/inigochoa/porg/releases/tag/0.1.0